+++ src/utils/__init__.py
@@ -4,6 +4,7 @@
 
 from .translations import translate_book_title, translate_author_name, get_book_alternative_title
 from .file_utils import safe_title, load_book_info
+from .affiliate_links import generate_affiliate_section
 
 __all__ = [
     'translate_book_title',
@@ -11,5 +12,6 @@ __all__ = [
     'get_book_alternative_title',
     'safe_title',
     'load_book_info',
+    'generate_affiliate_section',
 ]